+++ R/scoped_na_if.r
@@ -51,12 +51,10 @@ NULL
 #' @export
 
 na_if_all <- function(.tbl, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
-    fun       = dplyr::mutate_all,
+    fun       = "mutate_all",
     .tbl      = .tbl,
     arguments = list(...),
     arg_names = as.list(substitute(list(...)))
@@ -67,12 +65,10 @@ na_if_all <- function(.tbl, ...) {
 #' @export
 
 na_if_not_all <- function(.tbl, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
-    fun       = dplyr::mutate_all,
+    fun       = "mutate_all",
     .tbl      = .tbl,
     arguments = list(...),
     arg_names = as.list(substitute(list(...))),
@@ -84,12 +80,10 @@ na_if_not_all <- function(.tbl, ...) {
 #' @export
 
 na_if_at <- function(.tbl, .vars, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
-    fun       = dplyr::mutate_at,
+    fun       = "mutate_at",
     .tbl      = .tbl,
     .vars     = .vars,
     arguments = list(...),
@@ -101,12 +95,10 @@ na_if_at <- function(.tbl, .vars, ...) {
 #' @export
 
 na_if_not_at <- function(.tbl, .vars, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
-    fun       = dplyr::mutate_at,
+    fun       = "mutate_at",
     .tbl      = .tbl,
     .vars     = .vars,
     arguments = list(...),
@@ -119,12 +111,10 @@ na_if_not_at <- function(.tbl, .vars, ...) {
 #' @export
 
 na_if_if <- function(.tbl, .predicate, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
-    fun        = dplyr::mutate_if,
+    fun        = "mutate_if",
     .tbl       = .tbl,
     .predicate = .predicate,
     arguments  = list(...),
@@ -136,12 +126,10 @@ na_if_if <- function(.tbl, .predicate, ...) {
 #' @export
 
 na_if_not_if <- function(.tbl, .predicate, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
-    fun        = dplyr::mutate_if,
+    fun        = "mutate_if",
     .tbl       = .tbl,
     .predicate = .predicate,
     arguments  = list(...),