+++ NEWS.md
@@ -1,3 +1,11 @@
+# fauxnaif 0.6.1
+
+* Replace argument `input` with `x`.
+* Advance `fauxnaif::na_if()` and scoped `na_if_*()` functions from `deprecate_soft()` to `deprecate_warn()`.
+* Only warn `fauxnaif::na_if()` is deprecated if arguments could not be handled by `dplyr::na_if()`.
+* Added `pkgdown` site.
+* Remove `inst/CITATION`.
+
 # fauxnaif 0.6.0
 
 * No longer produce a warning when no replacements are made.
@@ -6,6 +14,7 @@
 
 # fauxnaif 0.5.6
 
+* Initial CRAN release.
 * Scoped na_if_*() functions now call an internal function
   fauxnaif:::scoped_na_if() to reduce repeated code.